Fix issue with calculations after '=' is hit
After '=' was hit, pressing another number would
concatenate it to the number displayed. In order to
avoid this, I added a variable that tracks the last
button pressed that updated the display. If that
button is the equals button, nothing happens when
numbers are pressed unless an operator is pressed before
hand.